This needs to be an int

+++ Adventure/ADVENT.C
@@ -22,9 +22,9 @@ int main(int argc, char** argv)
     szFilePath[0] = '\0';
 #else
     {
-        size_t i;
+        int i;
         strcpy_s(szFilePath, FILENAME_MAX, argv[0]);
-        for (i = strlen(szFilePath) - 1; i >= 0 && szFilePath[i] != '\\'; i--);
+        for (i = (int)strlen(szFilePath) - 1; i >= 0 && szFilePath[i] != '\\'; i--);
         szFilePath[i + 1] = '\0';
     }
 #endif
